+++ tests/test_catch_py311.py
@@ -176,3 +176,15 @@ def test_bare_raise_in_handler():
     assert excgrp.value is not middle_exc
     assert excgrp.value.__cause__ is first_exc
     assert excgrp.value.__context__ is first_exc
+
+
+def test_bare_reraise_from_naked_exception():
+    with pytest.raises(ExceptionGroup) as excgrp:
+        try:
+            raise KeyError("foo")
+        except* KeyError:
+            raise
+
+    assert len(excgrp.value.exceptions) == 1
+    assert isinstance(excgrp.value.exceptions[0], KeyError)
+    assert str(excgrp.value.exceptions[0]) == "'foo'"