+++ CHANGES.rst
@@ -8,6 +8,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added special monkeypatching if `Apport <https://github.com/canonical/apport>`_ has
   overridden ``sys.excepthook`` so it will format exception groups correctly
   (PR by John Litborn)
+- Fixed bare ``raise`` in a handler reraising the original naked exception rather than
+  an exception group which is what is raised when you do a ``raise`` in an ``except*``
+  handler
 
 **1.1.3**